* main: (213 commits)
  web: bump @types/node from 24.9.0 to 24.9.1 in /packages/prettier-config (#17617)
  web: bump @types/node from 22.15.19 to 24.9.1 in /web (#17618)
  web: bump knip from 5.66.1 to 5.66.2 in /web (#17619)
  translate: Updates for file web/xliff/en.xlf in pt_BR (#17639)
  core, web: update translations (#17643)
  website/docs: rel notes 2025.10: add 3 more integration guides (#17641)
  providers/proxy: drop headers with underscores (#17650)
  core: bump astral-sh/uv from 0.9.4 to 0.9.5 (#17645)
  web: bump style-mod from 4.1.2 to 4.1.3 in /web (#17647)
  core: bump github.com/getsentry/sentry-go from 0.36.0 to 0.36.1 (#17646)
  website/integrations: add terraform cloud (#17610)
  website/integrations: add zendesk (#17541)
  core: bump djangorestframework from 3.16.0 (our fork) to v3.16.1 (official package) (#16594)
  enterprise: add prometheus metrics for license usage and expiry (#17606)
  ci: link to next. for pre-release docs (#17634)
  web: sync web/package-lock.json (#17611)
  website/integrations: random fixes (#17631)
  website/docs: add short-lived certificate recommendation (#17628)
  core, web: update translations (#17605)
  translate: Updates for file locale/en/LC_MESSAGES/django.po in pt_BR (#17622)
  ...

+++ authentik/blueprints/v1/importer.py
@@ -72,13 +72,15 @@ from authentik.providers.oauth2.models import (
     DeviceToken,
     RefreshToken,
 )
+from authentik.providers.proxy.models import ProxySession
 from authentik.providers.rac.models import ConnectionToken
+from authentik.providers.saml.models import SAMLSession
 from authentik.providers.scim.models import SCIMProviderGroup, SCIMProviderUser
 from authentik.rbac.models import Role
 from authentik.sources.scim.models import SCIMSourceGroup, SCIMSourceUser
 from authentik.stages.authenticator_webauthn.models import WebAuthnDeviceType
 from authentik.stages.consent.models import UserConsent
-from authentik.tasks.models import Task
+from authentik.tasks.models import Task, TaskLog
 from authentik.tenants.models import Tenant
 
 # Context set when the serializer is created in a blueprint context
@@ -121,10 +123,12 @@ def excluded_models() -> list[type[Model]]:
         SCIMProviderUser,
         Tenant,
         Task,
+        TaskLog,
         ConnectionToken,
         AuthorizationCode,
         AccessToken,
         RefreshToken,
+        ProxySession,
         Reputation,
         WebAuthnDeviceType,
         SCIMSourceUser,
@@ -138,6 +142,7 @@ def excluded_models() -> list[type[Model]]:
         DeviceToken,
         StreamEvent,
         UserConsent,
+        SAMLSession,
         Message,
         GroupChannel,
     )
@@ -308,6 +313,7 @@ class Importer:
 
         serializer_kwargs = {}
         model_instance = existing_models.first()
+        override_serializer_instance = False
         if (
             not isinstance(model(), BaseMetaModel)
             and model_instance
@@ -336,11 +342,7 @@ class Importer:
                 model=model,
                 **cleanse_dict(updated_identifiers),
             )
-            model_instance = model()
-            # pk needs to be set on the model instance otherwise a new one will be generated
-            if "pk" in updated_identifiers:
-                model_instance.pk = updated_identifiers["pk"]
-            serializer_kwargs["instance"] = model_instance
+            override_serializer_instance = True
         try:
             full_data = self.__update_pks_for_attrs(entry.get_attrs(self._import))
         except ValueError as exc:
@@ -363,6 +365,12 @@ class Importer:
                 entry=entry,
                 serializer=serializer,
             ) from exc
+        if override_serializer_instance:
+            model_instance = model()
+            # pk needs to be set on the model instance otherwise a new one will be generated
+            if "pk" in updated_identifiers:
+                model_instance.pk = updated_identifiers["pk"]
+            serializer.instance = model_instance
         return serializer
 
     def _apply_permissions(self, instance: Model, entry: BlueprintEntry):
@@ -441,7 +449,7 @@ class Importer:
                 self._apply_permissions(instance, entry)
             elif state == BlueprintEntryDesiredState.ABSENT:
                 instance: Model | None = serializer.instance
-                if instance.pk:
+                if instance and instance.pk:
                     instance.delete()
                     self.logger.debug("Deleted model", mode=instance)
                     continue
